refactor(apple): use feature flags evaluated in JS
