+++ test/pack.test.ts
@@ -154,9 +154,9 @@ describe("npm pack", () => {
       "ios/app.mjs",
       "ios/assetsCatalog.mjs",
       "ios/entitlements.mjs",
+      "ios/features.mjs",
       "ios/infoPlist.mjs",
       "ios/localizations.mjs",
-      "ios/newArch.mjs",
       "ios/pod_helpers.rb",
       "ios/privacyManifest.mjs",
       "ios/test_app.rb",
