+++ test/test_pod_helpers.rb
@@ -17,66 +17,6 @@ class TestPodHelpers < Minitest::Test
     end
   end
 
-  def test_bridgeless_enabled?
-    ENV.delete('RCT_NEW_ARCH_ENABLED')
-
-    # Bridgeless mode is first publicly available in 0.73
-    available_version = v(0, 73, 0)
-
-    # Bridgeless mode is enabled by default starting with 0.74
-    default_version = v(0, 74, 0)
-
-    refute(bridgeless_enabled?({}, 0))
-    refute(bridgeless_enabled?({}, available_version))
-
-    options = { bridgeless_enabled: true, fabric_enabled: true }
-
-    refute(bridgeless_enabled?(options, v(0, 72, 999)))
-    assert(bridgeless_enabled?(options, available_version))
-
-    # Bridgeless mode is enabled by default starting with 0.74 unless opted-out of
-    assert(bridgeless_enabled?({ fabric_enabled: true }, default_version))
-    refute(bridgeless_enabled?({ bridgeless_enabled: false, fabric_enabled: true },
-                               default_version))
-
-    # `RCT_NEW_ARCH_ENABLED` does not enable bridgeless
-    ENV['RCT_NEW_ARCH_ENABLED'] = '1'
-
-    refute(bridgeless_enabled?({}, v(0, 72, 999)))
-    refute(bridgeless_enabled?({}, available_version))
-    assert(bridgeless_enabled?({}, default_version))
-    refute(bridgeless_enabled?({ bridgeless_enabled: false }, default_version))
-  end
-
-  def test_new_architecture_enabled?
-    ENV.delete('RCT_NEW_ARCH_ENABLED')
-
-    # New architecture is first publicly available in 0.68, but we'll require 0.71
-    available_version = v(0, 71, 0)
-
-    refute(new_architecture_enabled?({}, 0))
-    refute(new_architecture_enabled?({}, available_version))
-
-    # New architecture is first publicly available in 0.68, but we'll require 0.71
-    refute(new_architecture_enabled?({ fabric_enabled: true }, v(0, 70, 999)))
-    assert(new_architecture_enabled?({ fabric_enabled: true }, available_version))
-
-    # `RCT_NEW_ARCH_ENABLED` enables everything
-    ENV['RCT_NEW_ARCH_ENABLED'] = '1'
-
-    refute(new_architecture_enabled?({}, v(0, 70, 999)))
-    assert(new_architecture_enabled?({}, available_version))
-
-    # `RCT_NEW_ARCH_ENABLED` disables everything
-    ENV['RCT_NEW_ARCH_ENABLED'] = '0'
-
-    refute(new_architecture_enabled?({}, v(0, 70, 999)))
-    refute(new_architecture_enabled?({}, available_version))
-    refute(new_architecture_enabled?({ fabric_enabled: true }, available_version))
-
-    ENV.delete('RCT_NEW_ARCH_ENABLED')
-  end
-
   def test_use_hermes?
     options = { path: '../node_modules/react-native' }
